# openapi.yaml
openapi: 3.0.3
info:
  title: Tender API
  description: Tender API Backend Service
  version: 1.0.0
servers:
  - url: 'http://localhost:8080'
paths:
...
  /api/bids/new:
    post:
      summary: Create bid
      operationId: bidsCreate
      tags:
        - BIDS
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BidCreateRequest"
      responses:
        '200':
          description: Bid created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bid"
        '400':
          description: Bid created error
          content:
            application/json: { }
        '405':
          description: Method not allowed
          content:
            application/json: { }
        '403':
          description: Status Forbidden
          content:
            application/json: { }
        '500':
          description: Status Internal Server Error
          content:
            application/json: { }
...
  /api/bids/{tenderID}/list:
    get:
      summary: Fetch list bids by tender
      operationId: fetchListByTender
      tags:
        - BIDS
      parameters:
        - in: path
          name: tenderID
          schema:
            type: string
          required: true
          description: id тендера
        - in: query
          name: username
          required: true
          schema:
            type: string
          description: Имя ответственного за организацию тендера
      responses:
        '200':
          description: Bids fetch successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bid"
        '400':
          description: bids fetch error
          content:
            application/json: { }
        '405':
          description: Method not allowed
          content:
            application/json: { }
        '403':
          description: Status Forbidden
          content:
            application/json: { }
        '500':
          description: Status Internal Server Error
          content:
            application/json: { }
  /api/bids/my:
    get:
      summary: Fetch list bids by user
      operationId: fetchListByUser
      tags:
        - BIDS
      parameters:
        - in: query
          name: username
          required: true
          schema:
            type: string
          description: Имя ответственного за предложения
      responses:
        '200':
          description: Bids fetch successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bid"
        '400':
          description: bids fetch error
          content:
            application/json: { }
        '405':
          description: Method not allowed
          content:
            application/json: { }
        '403':
          description: Status Forbidden
          content:
            application/json: { }
        '500':
          description: Status Internal Server Error
          content:
            application/json: { }
  /api/bids/{tenderID}/reviews:
    get:
      summary: Fetch reviews for bid
      operationId: FetchReviews
      tags:
        - BIDS
      parameters:
        - in: path
          name: tenderID
          schema:
            type: string
          required: true
          description: id тендера
        - in: query
          name: username
          required: true
          schema:
            type: string
          description: Имя ответственного за организацию этого тендера
        - in: query
          name: organizationId
          required: true
          schema:
            type: string
          description: id ответственной организации этого тендера
        - in: query
          name: authorUsername
          required: true
          schema:
            type: string
          description: Автор, отзывы на предложения которого  нужно вывести
      responses:
        '200':
          description: Bids fetch feedback successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"
        '400':
          description: Bids fetch feedback error
          content:
            application/json: { }
        '405':
          description: Method not allowed
          content:
            application/json: { }
        '403':
          description: Status Forbidden
          content:
            application/json: { }
        '500':
          description: Status Internal Server Error
          content:
            application/json: { }
...
  /api/tenders/new:
    post:
      summary: tender create
      operationId: TenderCreate
      tags:
        - TENDERS
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenderCreateRequest"
      responses:
        '200':
          description: tender created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bid"
        '400':
          description: tender created error
          content:
            application/json: { }
        '405':
          description: Method not allowed
          content:
            application/json: { }
        '403':
          description: Status Forbidden
          content:
            application/json: { }
        '500':
          description: Status Internal Server Error
          content:
            application/json: { }
...
components:
  schemas:
    BidCreateRequest:
      title: "Запрос на создание предложения"
      type: object
      required:
        - name
        - description
        - tenderId
        - status
        - creatorUsername
        - organizationId
      properties:
        name:
          type: string
          format: text
          example: "Имя предложения"
        description:
          type: string
          format: text
          example: "Описание предложения"
        creatorUsername:
          type: string
          format: text
          example: "User1"
        tenderId:
          type: string
          format: string
          example: "550e8400-e29b-41d4-a716-446655440000"
        organizationId:
          type: string
          format: string
          example: "550e8400-e29b-41d4-a716-446655440000"
        status:
          type: string
          format: text
          enum: [ 'CREATED' ]
...
    Feedback:
      type: object
      title: Отзыв на предложение
      x-go-type-skip-optional-pointer: true
      required:
        - id
        - bid_id
        - description
        - responsible
        - created_at
      properties:
        id:
          type: string
          format: string
          example: "550e8400-e29b-41d4-a716-446655440000"
        bid_id:
          type: string
          format: text
          example: "550e8400-e29b-41d4-a716-446655440000"
        description:
          type: string
          format: text
          example: "Отзыв"
        responsible:
          type: string
          format: string
          example: "550e8400-e29b-41d4-a716-446655440000"
        created_at:
          type: string
          format: string
          example: "25.01.1994"
    Bid:
      type: object
      title: "Предложение"
      x-go-type-skip-optional-pointer: true
      required:
        - id
        - name
        - description
        - status
        - tenderId
        - version
        - responsible
      properties:
        id:
          type: string
          format: text
          example: "a6929720-17b0-45b4-bd54-76350fe5119b"
        name:
          type: string
          format: text
          example: "Имя предложения"
        description:
          type: string
          format: text
          example: "Описание предложения"
        status:
          type: string
          format: text
          enum: [ 'CREATED', 'PUBLISHED', 'CANCELED', 'APPROVED' ]
        tenderId:
          type: string
          format: string
          example: "550e8400-e29b-41d4-a716-446655440000"
        version:
          type: integer
          format: text
          example: 1
        responsible:
          type: string
          format: string
          example: "550e8400-e29b-41d4-a716-446655440000"
...
    TenderCreateRequest:
      title: "Запрос на создание тендера"
      type: object
      required:
        - name
        - description
        - serviceType
        - status
        - organizationId
        - creatorUsername
      properties:
        name:
          type: string
          format: text
          example: "Имя тендера"
        description:
          type: string
          format: text
          title: "Описание тендера"
        serviceType:
          type: string
          format: string
          title: "Категория тендера"
        status:
          type: string
          format: text
          enum: [ 'CREATED', 'PUBLISHED', 'CLOSED' ]
          title: "статус тендера"
        creatorUsername:
          type: string
          format: text
          title: "Имя ответственного за организацию - создатель тандера"
        organizationId:
          type: string
          format: string
          title: "id ответственной организации тандера"
          example: "550e8400-e29b-41d4-a716-446655440000"
...